Create yamlConverter.py
This is yaml conversion script. It was never quite completed, but should at least be able to perform a conversion between a YAML step file and the new Snooty step directive. It can also account for inheritance. It spits the new step(s) out in a steps dir.

To run this, youll need to have:
1) A step file to convert, as well as any step files it inherits from, in the same repo as the script
2) all required libraries (see the import statements at the top)
3) Python 3.9.1. Would recommend using pyenv for this